+++ src/algoritmos/ordenacao/InsertionSort.h
@@ -1,6 +1,6 @@
 #ifndef INSERTIONSORT_H
 #define INSERTIONSORT_H
 
-void insertionSort(unsigned *vet, unsigned size, unsigned long *numComp, unsigned long *numReg);
+void insertionSort(unsigned *vet, unsigned size, unsigned *numComp, unsigned *numReg);
 
 #endif // INSERTIONSORT_H
