+++ src/algoritmos/ordenacao/MergeSort.h
@@ -2,8 +2,8 @@
 #define MERGESORT_H
 #include <fstream>
 
-void mergeSort(unsigned *vet, unsigned esq, unsigned dir, unsigned long *numComp, unsigned long *numReg);
+void mergeSort(unsigned *vet, unsigned esq, unsigned dir, unsigned *numComp, unsigned *numReg);
 
-void merge(unsigned *vet, unsigned esq, unsigned med, unsigned dir, unsigned long *numComp, unsigned long *numReg);
+void merge(unsigned *vet, unsigned esq, unsigned med, unsigned dir, unsigned *numComp, unsigned *numReg);
 
 #endif // MERGESORT_H
