+++ src/main/cenario1.cpp
@@ -24,7 +24,7 @@ void fluxoCenario1(Review* conjunto, unsigned n, fstream &saida){
     for (unsigned versao = 0; versao < 2; versao++){
 
         // Métricas de desempenho
-        unsigned long numComparacoes = 0, numCopias = 0;
+        unsigned numComparacoes = 0, numCopias = 0;
 
         switch(versao){
