+++ src/main/cenario3.cpp
@@ -25,7 +25,7 @@ void fluxoCenario3(unsigned* ids, unsigned n, fstream &saida){
     for (unsigned versao = 0; versao < 5; versao++){
 
         // Métricas de desempenho
-        unsigned long numComparacoes = 0, numCopias = 0;
+        unsigned numComparacoes = 0, numCopias = 0;
 
         switch(versao){
